增加oss  接口
